+++ clients/client/src/misc.rs
@@ -1,7 +1,6 @@
 use instant::{Duration, Instant};
 
 pub struct UpdateTracker {
-    next_redraw_update: Instant,
     next_tick_update: Instant,
 }
 
@@ -15,24 +14,13 @@ impl UpdateTracker {
     pub fn new() -> Self {
         let now = Instant::now();
         UpdateTracker {
-            next_redraw_update: now,
             next_tick_update: now,
         }
     }
 
-    pub fn update(&mut self, redraw_ms: u64, tick_ms: u64) -> (bool, bool) {
-        let mut redraw_update = false;
+    pub fn update(&mut self, tick_period: Duration) -> bool {
         let mut tick_update = false;
         let now = Instant::now();
-        let redraw_period = Duration::from_millis(redraw_ms.max(1));
-        let tick_period = Duration::from_millis(tick_ms.max(1));
-
-        if now >= self.next_redraw_update {
-            redraw_update = true;
-            while self.next_redraw_update <= now {
-                self.next_redraw_update += redraw_period;
-            }
-        }
 
         if now >= self.next_tick_update {
             tick_update = true;
@@ -41,6 +29,6 @@ impl UpdateTracker {
             }
         }
 
-        (redraw_update, tick_update)
+        tick_update
     }
 }