+++ creator/src/actionlist.rs
@@ -24,6 +24,7 @@ impl ActionList {
             Box::new(crate::tools::copy_tile_id::CopyTileID::new()),
             Box::new(crate::tools::copy_vcode::CopyVCode::new()),
             Box::new(crate::tools::create_center_vertex::CreateCenterVertex::new()),
+            Box::new(crate::tools::create_campfire::CreateCampfire::new()),
             Box::new(crate::tools::create_fence::CreateFence::new()),
             Box::new(crate::tools::create_linedef::CreateLinedef::new()),
             Box::new(crate::tools::create_palisade::CreatePalisade::new()),