+++ creator/src/actions/mod.rs
@@ -7,6 +7,7 @@ pub mod clear_profile;
 pub mod clear_tile;
 pub mod copy_tile_id;
 pub mod copy_vcode;
+pub mod create_campfire;
 pub mod create_center_vertex;
 pub mod create_fence;
 pub mod create_linedef;